feat(storage): close Phase 1 open items — store_id BIGINT + dual-write A/B
Three items listed as 'open' on Phase 1:

1. store_id widened from INT/i32 to BIGINT/i64 end-to-end.
2. Dual-write trigger savings: measured with a 2x2 A/B.
3. ALTER EXTENSION upgrade path: intentionally deferred.

=== store_id widening ===

SQL:
- pg_mentat/sql/10_narrow_storage.sql: 9 narrow tables now declare
  store_id BIGINT NOT NULL DEFAULT 0 (was: INT).
- pg_mentat/src/lib.rs inline bootstrap: mentat.stores.store_id is
  BIGSERIAL PRIMARY KEY (was: SERIAL).

Rust:
- pg_mentat/src/functions/pull.rs: 13 store_id: i32 -> i64 parameter
  types; helper get_store_id_from_schema return and SPI Option types.
- pg_mentat/src/functions/transact.rs: 1 parameter, 1 helper return,
  1 Option<i32> -> Option<i64>.
- pg_mentat/src/functions/entity.rs: let store_id: i32 -> i64.
- pg_mentat/src/functions/query.rs: resolve_store_id return type,
  Option<i32> -> Option<i64>.
- pg_mentat/src/functions/time_travel.rs: get_store_id_for_schema
  return and Option<i32>.

Verified via information_schema.columns: every store_id column
reports data_type = 'bigint'. Smoke test passes. Read-path
benchmark unchanged (as expected for a column-type widening).

=== A/B benchmark ===

Method: git worktree against pre-Phase-1 (56b5565, dual-write
trigger ON) vs post-Phase-1 (b3a1f0d, trigger OFF). Two runs each.
Raw CSVs + README live at benchmarks/results/2026-05-09-phase1-ab/.

Results:

  | scale | pre-P1 avg | post-P1 avg | delta       |
  | 1k    | 4 953 ms   | 2 990 ms    | ~40% faster |
  | 10k   | 46 796 ms  | 33 972 ms   | ~27% faster |
  | 100k  | 342 495 ms | 358 512 ms  | within noise |

At 1k and 10k the directional signal is consistent across both
pairs: removing dual_write_datoms_trigger + 9-column CHECK +
8 wide-row indexes measurably reduces per-transaction overhead.

At 100k each load takes 5-7 minutes and laptop I/O / thermal
variance moves the number by more than the Phase-1 effect (pre-P1
run 2 at 314s is actually faster than post-P1 run 2 at 416s).
Two samples per condition cannot separate the trigger cost from
machine noise at this scale.

Conclusion: Phase 1 produces a measurable write speedup at
transactional workloads. Large-scale measurement needs a hermetic
benchmark box with many samples per condition — Phase 2 work.

BENCHMARKS.md extended with a 'Phase 1 A/B' section linking to the
raw data dir.

=== ALTER EXTENSION ===

Intentionally not tested. pg_mentat has no external deployments,
so there is no pre-existing installed schema to migrate from. The
upgrade script for the first real transition will be written and
tested when that first deployment happens.

=== Docs ===

- STATUS.md: store_id known-limitation line removed (now load-bearing
  bigint). 'No pg_upgrade testing' line rewritten to reflect the
  intentional deferral.
- ROADMAP.md: Phase 1 Open follow-ups section replaced with a
  resolved/deferred breakdown.

Phase 1 is now fully closed.

+++ pg_mentat/sql/10_narrow_storage.sql
@@ -14,7 +14,7 @@
 -- ---------------------------------------------------------------------------
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_ref_new (
-    store_id INT     NOT NULL DEFAULT 0,
+    store_id BIGINT  NOT NULL DEFAULT 0,
     e        BIGINT  NOT NULL,
     a        BIGINT  NOT NULL,
     v        BIGINT  NOT NULL,
@@ -24,56 +24,56 @@ CREATE TABLE IF NOT EXISTS mentat.datoms_ref_new (
 ) WITH (fillfactor = 90);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_long_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v BIGINT NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
 ) WITH (fillfactor = 90);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_text_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v TEXT NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
 ) WITH (fillfactor = 85, toast_tuple_target = 8160);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_double_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v DOUBLE PRECISION NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
 ) WITH (fillfactor = 90);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_instant_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v TIMESTAMPTZ NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
 ) WITH (fillfactor = 90);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_keyword_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v TEXT NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
 ) WITH (fillfactor = 90);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_uuid_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v UUID NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
 ) WITH (fillfactor = 90);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_bytes_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v BYTEA NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
 ) WITH (fillfactor = 85, toast_tuple_target = 8160);
 
 CREATE TABLE IF NOT EXISTS mentat.datoms_boolean_new (
-    store_id INT NOT NULL DEFAULT 0,
+    store_id BIGINT NOT NULL DEFAULT 0,
     e  BIGINT NOT NULL, a BIGINT NOT NULL, v BOOLEAN NOT NULL,
     tx BIGINT NOT NULL, added BOOLEAN NOT NULL DEFAULT TRUE,
     PRIMARY KEY (store_id, e, a, tx)
